move pkg/models/datashards -> pkg/models/topology/shards.go
